services/watchlist-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
asyncpg==0.29.0
PyJWT==2.8.0
structlog==24.1.0
httpx==0.26.0
